Added a few, fixed a few

+++ ft_strlen.c
@@ -11,10 +11,11 @@
 /* ************************************************************************** */
 
 #include "libft.h"
+#include <stddef.h>
 
-unsigned long int	ft_strlen(const char *str)
+size_t ft_strlen(const char *str)
 {
-	unsigned long int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
