+++ libft.h
@@ -12,13 +12,14 @@
 
 #ifndef LIBFT_H
 # define LIBFT_H
+# include <stddef.h>
 
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
 int					ft_isalnum(int c);
 int					ft_isprint(int c);
 int					ft_isascii(int c);
-unsigned long int	ft_strlen(const char *str);
+size_t				ft_strlen(const char *str);
 int					ft_strlcat(char *dst, const char *src, int size);
 int					ft_strlcpy(char *dest, const char *src, int size);
 int					ft_toupper(int ch);
